+++ src/process_utils/precision.c
@@ -6,10 +6,11 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 21:09:53 by kamitsui          #+#    #+#             */
-/*   Updated: 2023/03/24 10:53:23 by kamitsui         ###   ########.fr       */
+/*   Updated: 2023/03/25 16:20:34 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
 #include "ft_printf.h"
 #include "process.h"
 #include "libft.h"
@@ -17,7 +18,7 @@
 int	precision(const char *input, t_sm *machine)
 {
 	static char	*str_prec;
-	int			size;
+	size_t		size;
 
 	str_prec = FIELD_PREC;
 	size = 0;
@@ -26,10 +27,8 @@ int	precision(const char *input, t_sm *machine)
 		machine->flag |= BIT_PREC;
 		size++;
 		if (ft_isdigit(input[size]) != TRUE)
-				return (size);
+			return (size);
 		size += get_field_nbr(&input[size], machine);
 	}
 	return (size);
 }
-//		if (size == -1)
-//			return (size);
